requirements.txt
streamlit
google-genai
google-cloud-logging  # <--- Make sure this line exists!